Move internal.check_lock to its own module
